ExampleJavelin merge & improvements
